Typescript conversion
another BIG, BIG change
* converted all code to Typescript
* initial bundling with webpack to `gridstack.all.js` for browser inclusion
* added index.ts to include all modules and dependencies for build
* part-2 of gridstack #1084

TODO:
* modified float.html to test new code. still not running (missing `GridStack` export)
* need to use tsc to output for typescript native files ?
