+++ .npmignore
@@ -12,12 +12,15 @@ yarn.lock
 .npmignore
 .travis.yml
 .eslint*
+.vscode
 Gruntfile.js
 ISSUE_TEMPLATE.md
 PULL_REQUEST_TEMPLATE.md
 freelancer
 karma.conf.js
 protractor.conf.js
+tsconfig.json
+webpack.config.js
 
 ## From .gitignore:
 *.log
