+++ demo/float.html
@@ -7,10 +7,13 @@
   <title>Float grid demo</title>
 
   <link rel="stylesheet" href="demo.css"/>
+  <script src="../dist/gridstack.all.js"></script>
+  <!--
   <script src="../src/jquery.js"></script>
   <script src="../src/gridstack.js"></script>
   <script src="../src/jquery-ui.js"></script>
   <script src="../src/gridstack.jQueryUI.js"></script>
+  -->
   </style>
 </head>
 <body>
