+++ src/gridstack-extra.scss
@@ -1,5 +1,5 @@
 /*!
- * gridstack 1.1.0-dev extra CSS for [2-11] columns (non default)
+ * gridstack 2.0.0-rc extra CSS for [2-11] columns (non default)
  * https://gridstackjs.com/
  * (c) 2014-2020  Alain Dumesny, Dylan Weiss, Pavel Reznikov
  * gridstack.js may be freely distributed under the MIT license.
