+++ src/gridstack.scss
@@ -1,5 +1,5 @@
 /*!
- * required gridstack 1.1.0-dev CSS for default 12 and 1 columnMode size. Use gridstack-extra.css for others
+ * required gridstack 2.0.0-rc CSS for default 12 and 1 columnMode size. Use gridstack-extra.css for others
  * https://gridstackjs.com/
  * (c) 2014-2020 Alain Dumesny, Dylan Weiss, Pavel Reznikov
  * gridstack.js may be freely distributed under the MIT license.
