+++ src/index.ts
@@ -1,0 +1,13 @@
+// index.ts 2.0.0-rc @preserve
+
+/**
+ * The package's root file.  Exports all exported classes.
+ */
+import './gridstack-poly.js';
+
+export * from './types';
+export * from './utils';
+export * from './gridstack-engine';
+export * from './gridstack-dragdrop-plugin';
+export * from './gridstack';
+export * from './jqueryui-gridstack-dragdrop-plugin';
